Clean up bundler
